add and fix errors
