+++ Geekco-Store2.0/public/stylesheets/registro.css
@@ -29,9 +29,20 @@
     padding-bottom: 10px;
     font-family: 'Pixelify Sans', sans-serif;
 }
-.register-form h2, .register-form p {
+.register-form h2, .register-form .info-text-form p {
     text-align: center;
 }
+.text-error {
+    animation: text-error 3500ms infinite;
+    font-size: 12px;
+  }
+  
+  @keyframes text-error {
+    0% {color: var(--color-turquoise);}
+    50% {color: white;}
+    100% {color: var(--color-turquoise);}
+  
+  }
 .register-form p  {
     font-size: 18px;
 }
