+++ Geekco-Store2.0/src/views/error.ejs
@@ -1,7 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-    <%-include("./partials/head.ejs", {title: "Error!"})%>
-<body>
-    <p>Error!!!</p>
-</body>
-</html>
+<h1><%= message %></h1>
+<h2><%= error.status %></h2>
+<pre><%= error.stack %></pre>
