+++ Geekco-Store2.0/src/views/index.ejs
@@ -38,6 +38,5 @@
           </div>
         </main>
         <%-include("./partials/footer.ejs")%>
-        <script src="./js/main.js"></script>   
 </body>
 </html>
