+++ Geekco-Store2.0/src/views/partials/header.ejs
@@ -44,5 +44,6 @@
             
         </nav>
         <div class="header__gradient"></div>
+
     </header>
-    <script src="../../../public/js/main.js"></script>   
+    <script src="/js/main.js"></script>
