+++ Geekco-Store2.0/src/views/users/register.ejs
@@ -20,50 +20,71 @@
                 <label for="first_name">Nombre</label>
                 <input type="text" name="first_name" id="name-user" value="<%=locals.old&&old.first_name? old.first_name:`` %>">
                 <span class="asterisk-form">*</span>
+                <div class="text-error">
                 <% if(locals.errors && errors.first_name){ %>
-                    <p style="text-align: center;margin: auto;color: red;font-weight: 500;"></p><%= errors.first_name.msg  %></p>
+                    <p><%= errors.first_name.msg  %></p>
                     <% } %>
+                </div>
             </div>
             <div class="input-data">
                 <label for="last_name">Apellido</label>
                 <input type="text" name="last_name" id="lastname-user"value="<%= locals.old&&old.last_name? old.last_name:``  %>">
                 <span class="asterisk-form">*</span>
+                <div class="text-error">
                 <% if(locals.errors && errors.last_name){ %>
-                    <p style="text-align: center;margin: auto;color: red;font-weight: 500;"></p><%= errors.last_name.msg  %></p>
+                    <p><%= errors.last_name.msg  %></p>
                     <% } %>
+                </div>
             </div>
             <div class="input-data">
                 <label for="email">Dirección de correo eléctronico</label>
                 <input type="email" name="email" id="email-user" value="<%= locals.old&&old.email? old.email:``  %>">
                 <span class="asterisk-form">*</span>
+                <div class="text-error">
                 <% if(locals.errors && errors.email){ %>
-                    <p style="text-align: center;margin: auto;color: red;font-weight: 500;"></p><%= errors.email.msg  %></p>
+                    <p><%= errors.email.msg  %></p>
                     <% } %>
+                </div>
             </div>
             <div class="input-data">
                 <label for="adress">Domicilio</label>
                 <input type="text" name="adress" id="direction-user" value="<%= locals.old&&old.adress? old.adress:``  %>">
+                <div class="text-error">
                 <% if(locals.errors && errors.adress){ %>
-                    <p style="text-align: center;margin: auto;color: red;font-weight: 500;"></p><%= errors.address.msg  %></p>
+                    <p><%= errors.address.msg  %></p>
                     <% } %>
+                </div>
             </div>
             <div class="input-data">
                 <label for="password">Contraseña</label>
                 <input type="password" name="password" id="password-user">
                 <span class="asterisk-form">*</span>
+                <div class="text-error">
                 <% if(locals.errors && errors.password){ %>
-                    <p style="text-align: center;margin: auto;color: red;font-weight: 500;"></p><%= errors.password %></p>
+                    <p><%= errors.password.msg %></p>
                     <% } %>
+                </div>
             </div>
             <div class="input-data">
-                <label for="confirm-password">Confirmar contraseña</label>
-                <input type="password" name="confirm-password" id="confirm-password-user">
+                <label for="confirm_password">Confirmar contraseña</label>
+                <input type="password" name="confirm_password" id="confirm_password">
                 <span class="asterisk-form">*</span>
+                <div class="text-error">
+                <% if(locals.errors && errors.confirm_password){ %>
+                    <p><%= errors.confirm_password.msg %></p>
+                    <% } %>
+                </div>
             </div>
+
             <div class="input-data">
-                <label for="imagen">selecciona una imagen para el usuario</label>
+                <label for="imagen">Selecciona una imagen para el usuario</label>
                 <input type="file" name="image" id="imagen">
                 <span class="asterisk-form">*</span>
+                <div class="text-error">
+                <% if(locals.errors && errors.image){ %>
+                    <p><%= errors.image.msg %></p>
+                    <% } %>
+                </div>
             </div>
             <div class="info-text-form">
                 <p>Tus datos personales se utilizarán para procesar tu pedido, mejorar tu experiencia en esta web, gestionar el acceso a tu cuenta y otros propósitos.</p>
